feat(patch): applied centre, grid, gridnav and listfullwidth
Combined with previous 'managed' patch and a rule in dwm/config.def.h,
makes demenu appear as interactive window in the centre of the screen.
Centre patch is used in this case to provide an initial window size —
smaller than dmenu would otherwise be on its own — when spawning as a
floating window.

List items in a dmenu will not reflow when the window is resized. But
at least text doesn't get cut of either like what happen in st.